requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
mailjet-rest==1.3.4
python-dotenv==0.19.2
requests==2.26.0
supermarktconnector==0.5
urllib3==1.26.7
